refactor: Consolidate HTTP and file handling by removing obsolete classes and updating namespaces and moving to dedecated folder
