+++ src/Facades/AsyncHttp.php
@@ -3,7 +3,7 @@
 namespace Rcalicdan\FiberAsync\Facades;
 
 use Rcalicdan\FiberAsync\Contracts\PromiseInterface;
-use Rcalicdan\FiberAsync\Handlers\AsyncOperations\HttpHandler;
+use Rcalicdan\FiberAsync\Handlers\Http\HttpHandler;
 use Rcalicdan\FiberAsync\Http\Request;
 
 /**
